requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-import-export==3.3.6
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.4
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
